+++ src/Resources/ResourceManager/Implementation/Features/UnregisterAzureProviderPreviewFeatureCmdlet.cs
@@ -15,14 +15,13 @@
 namespace Microsoft.Azure.Commands.ResourceManager.Cmdlets.Implementation
 {
     using System.Management.Automation;
-    using Microsoft.Azure.Commands.ResourceManager.Cmdlets.SdkModels;
     using ProjectResources = Microsoft.Azure.Commands.ResourceManager.Cmdlets.Properties.Resources;
 
     /// <summary>
-    /// Creates feature registration.
+    /// Unregisters feature registration.
     /// </summary>
-    [Cmdlet("New", ResourceManager.Common.AzureRMConstants.AzureRMPrefix + "FeatureRegistration", SupportsShouldProcess = true), OutputType(typeof(PSSubscriptionFeatureRegistration))]
-    public class NewAzureFeatureRegistrationCmdlet : ProviderFeatureCmdletBase
+    [Cmdlet("Unregister", ResourceManager.Common.AzureRMConstants.AzureRMPrefix + "ProviderPreviewFeature", SupportsShouldProcess = true), OutputType(typeof(bool))]
+    public class UnregisterAzureProviderPreviewFeatureCmdlet : ProviderFeatureCmdletBase
     {
         /// <summary>
         /// Gets or sets the provider name
@@ -39,15 +38,29 @@ namespace Microsoft.Azure.Commands.ResourceManager.Cmdlets.Implementation
         [ValidateNotNullOrEmpty]
         public string ProviderNamespace { get; set; }
 
+        [Parameter(Mandatory = false)]
+        /// <summary>
+        /// Gets or sets the pass thru.
+        /// </summary>
+        public SwitchParameter PassThru { get; set; }
+
         /// <summary>
         /// Executes the cmdlet
         /// </summary>
         public override void ExecuteCmdlet()
         {
             this.ConfirmAction(
-                processMessage: ProjectResources.NewFeatureRegistrationMessage,
+                processMessage: ProjectResources.RemoveFeatureRegistrationMessage,
                 target: this.ProviderNamespace,
-                action: () => this.WriteObject(this.ProviderFeatureClient.CreateFeatureRegistration(providerName: this.ProviderNamespace, featureName: this.Name)));
+                action: () =>
+                {
+                    this.ProviderFeatureClient.DeleteFeatureRegistration(providerName: this.ProviderNamespace, featureName: this.Name);
+
+                    if (this.PassThru.IsPresent)
+                    {
+                        this.WriteObject(true);
+                    }
+                });
         }
     }
 }